Rework store to remove api calls
Store will no longer populate with an API call at the first loading of
the app.  The store will simply create a store with null values for all
of the various keys that will be needed.  The store will be populated
for the first time during the mounting of the App component