Remove store initialization
Store will no longer be intialized before mounting the app to the DOM.
With the addition of dynamic url routing... the App will now wait to
read the url parameters before deciding which product to populate the
store with.  Instead a blank store will be created and wrapped around
the app in this phase.